+++ android/app/build.gradle.kts
@@ -61,6 +61,7 @@ dependencies {
     implementation("androidx.compose.material:material-icons-extended")
     implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.6.2")
     implementation("androidx.lifecycle:lifecycle-runtime-compose:2.6.2")
+    implementation("androidx.navigation:navigation-compose:2.7.6")
 
     implementation("com.github.alorma:compose-settings-ui-m3:1.0.3")
     implementation("com.github.alorma:compose-settings-storage-preferences:1.0.3")
